Added animations to faq window and sponsors
