+++ apps/site/src/app/(home)/page.tsx
@@ -26,7 +26,7 @@ const Home = ({ searchParams }: HomeProps) => {
 			<Landing />
 			<About />
 			{/* <Countdown /> */}
-			{/* <Sponsors /> */}
+			<Sponsors />
 			{/* <Clubs /> */}
 			<FAQ />
 			<BackToTop />
